// src/app/shared/enums/DatingEnum.ts
export enum DatingEnum {
  anytime = 0,
  tomorrow = 1,
  today = 2,
  yesterday = 3,
  onOrAfter = 4,
  on = 5,
  OnOrBefore = 6,
  between = 7,
  nextWeek = 8,
  thisWeek = 9,
  lastWeek = 10,
  nextMonth = 11,
  thisMonth = 12,
  lastMonth = 13,
  nextYear = 14,
  thisYear = 15,
  lastYear = 16
}